requirements.txt
asgiref==3.7.2
chardet==5.1.0
click==8.1.3
colorama==0.4.6
Django==4.2.2
gensim==4.3.1
gunicorn==20.1.0
h11==0.14.0
joblib==1.2.0
lxml==4.9.2
nltk==3.8.1
numpy==1.24.3
python-docx==0.8.11
python-dotenv==1.0.0
regex==2023.6.3
scikit-learn==1.2.2
scipy==1.10.1
smart-open==6.3.0
sqlparse==0.4.4
threadpoolctl==3.1.0
tqdm==4.65.0
typing-extensions==4.6.3
tzdata==2023.3
uvicorn==0.22.0
